+++ lib/util/extension/ext_number_rounding.dart
@@ -1,12 +1,12 @@
 extension NumberRounding on double {
 
-  int roundTo(int number) {
+  double roundTo(int number) {
 
     if (this == null) {
       return 0;
     }
 
-    return double.parse((this).toStringAsFixed(number)).toInt();
+    return double.parse((this).toStringAsFixed(number));
   } 
 
 }
